Extract SosClient interface and move implementation to package web
